+++ c/hw_drivers/spi/spi_i94xxx/src/spi_i94xxx.h
@@ -18,10 +18,10 @@ struct spi_i94xxx_cfg_t {
 	struct dev_desc_t *   src_clock;
 	uint32_t   clk_freq;
 	size_t     base_address;
-	uint8_t    CLK_pin;
-	uint8_t    SS_pin;
-	uint8_t    MISO_pin;
-	uint8_t    MOSI_pin;
+	uint32_t    CLK_pin;
+	uint32_t    SS_pin;
+	uint32_t    MISO_pin;
+	uint32_t    MOSI_pin;
 	uint8_t    data_width;
 	uint8_t    rx_fifo_threshold;
 	uint8_t    tx_fifo_threshold;
